this is what i got so far
there is an issue where it doesn't draw the path to the 3rd ability. I
think i know why, we need to figure out how to make so we can split it
up into segments in the sense that when it finds one ability, save that
portion of the path in analysis, then when it finds the next ability
save that path to analysis and so on.  feel free to try something
different to get it to work